refactor(cli_commands): Tweak implementation + add comments
This will tweak the implementations of both CLI commands to clean them up slightly, as well as add documentation for them, commenting how these processes work.

BREAKING CHANGE: Note: While the `lib.rs` module for the `minigrep` binary exists in a very limited scope, the tweaks to the naming of the public interface of the `Config` struct, would still constitute a breaking change in the case they'd seen implementation.