feat(fizzbuzz): Implement `fizzbuzz` factory + update command implementation
This adds an function which constructs an iterator that returns the `FizzBuzz` game sequence.

Given I'd already written an implementation of this as a command binary, this also elects to rewite it's implementation, to utilise this new factory.

Additonally, I've made some minor adjustment to the documentation, hiding the imports used in examples which exclusively import themselves.