feat(input): Rework input handling tools
I once against came to the decision I was unhappy with my implementation of custom input handling, deciding to once again to rework my tools, utilise a newtype pattern instead of traits, due to the limited applicability of their designs outside of the specific context.

As such, the list of methods for the type have been further cut down, to simply two: One of which reads a line of console input to a new string. The other, reads lines of input, to the specified range. Both of these utilise the `std::io::Result` alias, propagating any errors resulting from the process.

With the redesign of the current system, I also reworked tests and examples dependant on the previous designs to instead utilise the new system.

BREAKING CHANGE: With the redesign of my system, I removed the previous implementation of the system, in order to deny me from using them.