feat(my_rusttools): Flatten module paths
Given the minimal contents of the `gcacher` and `input` modules, rather than having to specify their now underlying modules to access them, I chose to flatten the file paths to their main structs.

BREAKING CHANGE: With the change to paths to these tools, this required flattening the paths used within other areas of the workspace.