feat(reitertools): Initial library implementation
Due to my disagreement with implementation aspects of the `itertools` library, I decided to make my own design implementations. In turn, this is the result.

Currently, the library features only minimal implementaitons, notably, my prefered implementation of a method to count the number of times unique elements are taken from an iterator, along with methods based on it, as well as to tally a single item.

More notably, a meta-iterator struct, for defining bespoke advancement behaviour based on the inner iterator, such as batching processes or struct specific methods. This also results in an implementation to check if the elements of an iterator are a palindrome, by batching with iteration from both ends of an iterator.