feat(is_palindrome): Implement palindrome checker
With the creation of my `reitertools` library, I realised it allowed for an effective method to check if a collection was a palindrome.

Specifically in the case of strings, this required a somewhat bespoke implementation, and as such, I forked the developement branch of my `my_rusttools` library, to merge the unfinished `reitertools` library into it, and implement a function to check whether a string is a palindrome.