refactor(factories::iter_factories): Function reworks
Previously, the `sieve_primes()` factory function, used two nested `for` loops to mark non-prime indexes.

This reworks the function, to match for cases where the upper bound `0` or `1` (neither of which can be primes), and uses `fold()` with a stepped `RangeFrom`, to iterate for every odd index for the length of the marker `Vec`, and mark every multiple of the current index as not a prime, before returning the resulting `Vec`.

Previously, the adapter `fizzbuzz()` created, used an equality check between the values of the `fizzbuzz` iterator, to determine whether the current index was not divisible by `3` or `5`.

This reworks the adaptor to instead use pattern matching, with a match of the tuple contents against empty strings, returning the current index as a string, otherwise, destructuring the tuple, appending the string values and returning them. This change was made to make the design easier to potentially extend in future with more values.