refactor(traits::humainze_bool): Lifetime correction
Previously, the signitures of the methods of the `HumanizeBooleans` trait, used named lifetimes for it's return values.

This corrects them to instead use `'static` lifetimes.

BREAKING CHANGE: While this change will technically break implementation cases, due to changing the signiture of the traits methods, they were designed with the intent to referance static values, as such, making a `'static` lifetime more appropriate.