refactor(factories::iter_factories): Improve comments + Replace `skip()`
In sharing my code, I reaslised a good deal of my comments were redundant, or could be more concise or effectively structured. There was also a redundant element in one of my documentaiton examples, which is also removed.

Previously, the final adaptor created by the function, uased `enumerate()`, `skip()`, `filter()` and `map()` to return the indexes to pick primes from, drop the first two, filter the given prime values, then drop the marker element from the tuple.

This drops all three of the last three adaptors in favour of the `filter_map()` adaptor, with an arms to drop the `0` and `1` index, effectively skipping them, explicitly return index `2`, check if the index isn't an even number, and is still marked as a prime number, and otherwise drops the value.

Overally, this should significantly minimise the adaptors used, in a manner which is also more readable.