refactor(factories::iter_factories): Explicit even index arm
Previously, the adaptor produced by the `sieve_primes()` function, used the condifiton, `x % 2 != 0 && y` to include indexes which aren't odd, and still marked as primes.

This adds a matching arm using the condition `x % 2 == 0`, to denote explicitly, in a more typical manner, that even values are being dropped by the adapter.

To coincide with this change, comments on the function have also been adjusted accordingly.