feat(gcacher::GCacher): Clean up docs + getters + Add Clone
With my revist to some of my old work, I've taken the time to do some minor clean-up on the implementation, and documentation on my `GCacher` struct, consequently, also introducing a new dependancy to the crate.

Previously, I removed the getter methods for the `instancer` field, due to the fact I made it public due to being a closure, which cannot be meaningfully mutated. Additionally, `cache` has an explicit getter, with overly verbose documentation.

The getters for `GCacher`s fields, are not implemented and documented by marcros from the `getset` crate.

Additionally, the `try_reserve()` method has now been implemented for the `GCacher` struct, now it has been stablised on the `HashMap` struct.