feat(factories::iter_factories): Repeat interval functions + Fizzbuzz refactor
Following posting a challenge to implement a version of the game FizzBuzz, without using a division operation, to the Rust community Discord server, some proposed solutions introduced me to functions I was previously unaware of, which would allow me to programatically generate the sequences I'd created for the `fizzbuzz()` function, rather than having to hardcode them.

As such, this now implements the `repeat_interval()` and `repeat_interval_with()` functions, to produce cyclic interval iterators, which are in turn used to refatcor the `fizzbuzz()` function to be more concise.