refactor(pigify): Clean up implementaiton
Having been implemented early in my time learning Rust, its implementation very much had it's hallmarks of inexperience. Previously, the function worked on a single function scope, iterating over the words in the referanced string using a `for` loop, in combination with `if` statements to determine the graphere to append to the translated string.

This implementaiton instead takes advantage of the `fold()` method, to iterate over every word, switching to a guard clause to exit from the iteration early when an item isn't a word. Following this, a guarded `match` statement, is used to define whether the leading grapheme is returned as the leading, or trainling, value of the resulting string for the iteration by whether the value is a consonant. Finally, the remainder of the item, and resulting values from the `match` statement, are formatted into a `String`, which is trimmed if nessacary, before being appended to the builder string, which is returned.