refactor(factories::iter_factories): `successors()` + Type aliases refactor
Previously, the `sieve_primes()` was implemented using `RangeFrom`, `skip()` and `step_by()` iterator adapter, to iterate over the indexes for primes. Due to utilising the `advance_by()` method to eagerly advance an iterator, this would potentially result in a lot of unnecessary instructions, where a simple mathematical sequence can instead be used.

As such, this changes the implementation to instead use the `successors()` function to create these sequences, including using a `fold()` method over the multiples of each index and mark them as not primes. This is alongside a guard clause for the closure for the odd indexes fold, to check the state of the indexes before iterating over the multiple of the index, avoiding doing so if unnecessary.

Additionally, the return types of the functions in the module tended to be `impl Iterator`, which had the unfortunate issue of not resulting in concrete return types, despite having consistant return types. Due to this unfortunate implementation fact, I'd always been unhappy, due to this resulting in only being able to use the implementation tied to the `Iterator` trait, without listing every trait the return type used at length, but I struggled to exchange this implementaiton for type aliases, due to lacking the knowledge to correctly structure and apply them. This finally applies this change, due to working out how to correctly structure aliases and their returns.

Finally, this also adds a test to check primes up to 10000.