refactor(pigify): Improve control flow conciseness
Previously, the `pigify()` function used a `match` statement with a match guard for testing whether the first grapheme of the iteration string, was a vowel, using a nested closure, based on the `contains()` method on a `&str` of vowels.

This swaps the `&str` of vowels, for a `&[char]` of vowels, along with swapping the control flow from using a `match` statement, to a `map()` method with a ternary `if` statement, using the `contains()` method and the `VOWELS` variable, due to the `Pattern` trait being implemented for the `&[char]` type, and an `expect()` to unwrap the resulting value, resulting in more concise control flow.