feat(pigify): Add pre-allocation + Misc tests
Previously, the `pigify()` function would simply allocate an empty string to append the resulting Pig Latin `String` to.

Given the resulting string should scale somewhat consistently with the number of words in the string plus it's length, this meant an expected capacity could be calculated. As such, this updates the implementation to include that calculation and resulting allocation, alongside changing the return type to allow for overflow errors, as a formality.

Additionally, this creates a new file for misc. integrsation tests, notably adding a test for `pigify()` against the quick brown fox phrase.

BREAKING CHANGE: Due to the including for cases of overflow errors, this changes the signiture for the `pigify()` command.