refactor(factories::iter_factories): Safety + Signiture improvements
Previously, the `sieve_primes()` function implemented no checks regarding the safety of the mathematical operations it performed. In turn, adding safety checks as a formality, constitues the bulk of the changes in this commit, despite the fact allocation of a `Vec` of the size to cause such errors, isn't possible without allocation errors to begin with.

Additionally, this makes a minor change to the `FizzBuzz` type alias, making it less generalised, but also more concise in regards to the signiture of the `fizzbuzz()` function.