refactor(my_rusttools::sieve_primes): Implementation revision
With my previous commit to this file, it seems pretty clear I believed the performance of the `Skip` and `StepBy` adaptors, would be poor, due to assuming they would use the default implementations of `nth()` and `advance_by()`.

Upon actually investigating the implementations of the types my original implementation utilised, I reaslised they didn't use the default implementations, and in-fact, effectively utilised the implemtation I put together myself, in particular, the implementation of `IterMut`.

As such, this reverts the inner implementation to use `Skip` and `StepBy` adaptors, and retains the outer use of `successors()`, to step over the odd indexes, whilst also checking the power of the index, doesn't overflow, or exceed the specified upper bound. Alongside this, `fold()` is no longer used by the implementation, in favour of reverting to the use of a mutable variable for the `Vec` of markers, for the sake of simplicity, as well as the addition of iteration over the first two markers, to mark them as not prime.

Additionally, the structure of the `match` statement used by the resulting iterator from `sieve_primes()` has been improved, checking the marker on the index is `false` on the first arm to filter out the most common case, followed by checking the index is odd, with a check for the index `2`, being the least common accountable case.

Otherwise this makes some minor changes to wording in some documentation.