refactor(my_rusttools::fizzbuzz): Implement overflow protection
Previously, `fizzbuzz()` produced an iterator which provided no bounding protection, due to my assumption that the release behaviour for overflowing an unbounded range, may not be panicing. Upon inspecting the source code, I realised this assumption was not the case, meaning not providing such protection, makes no sense.

As such, this commit switches to using a `Successors` iterator, using `checked_add()` operations, to calculate the indexes successor. Additionally, this elicits a change to the signiture of the type-alias used by the function, to add an additional generic parameter.

BREAKING CHANGE: Due to the change in the implementation of `fizzbuzz()`, the addition of another generic parameter to the signiture of the type-alias, `FizzBuzz`, was required.