test(my_rusttools::factory_tests): Improve clarity + Decoupling
This commit should improve the clarity of some of the implementation tests, as well as decoupling unrelated tests, notabling `repeat_foo()`, `repeat_from_bar()`, and `repeat_ab()`. Additionally, with the change to the implementation of `fizzbuzz()`, the `fizzbuzz_is_infinite()` test is removed with this commit, due to the idea of the test, being inherrently flawed.