feat(my_rusttools::StdinExtended): Refine `read_lines()`
Ever since I concived of the `read_lines()` method, I've been unhappy with the implementation, and whilst this remains true, following learning of the `Lines` iterator, but this attempts to somewhat refine the implementation provided by this method.

Notably, rather than continutally appending lines of input to an single string, this implementation collects the lines into a `Vec` until the method is complete, wherein all the lines are appended with a single method call. Whilst doing nothing to guarantee the safety of the operation, it is significantly less wasteful than the previous approach of continually reconstructing the string with each iteration.

BREAKING CHANGE: The new implementation of `read_lines()` is now guaranteed to propagate any IO errors, as such, eliciting a change to the method's signiture, entirely removing the `err_notif` variable. Alongside this, the signiture of the `notif` variable has also been changed, to now provide exclusively the current line count, due to not posessing access to the resulting collection of input, due to the structure of the implementation.