refactor(my_rusttools::humanize_bool) Streamline Implementations
With the update to Rust 1.62.0, and the stabilisation of the  method for s, I've been looking for implementations which could be streamlined using it.

This commit is part of my effort to make use of this newly stabilised feature, in this case, by streamlining the implementations of the  trait for the  type, using  and  methods, instead of  statements.