refactor(my_rusttools::pigify): Switch to `Cow` and `collect()`
Previously, the `pigify()` function, estimated the resulting length of
the string it would produce (including propagating an error variant if
it might've exceeded available heap capacity). This was used with
`String::with_capacity` and a `fold()` over the function argument,
appending resulting values from the pigifying process,
before returning the string.

Using `fold()` like this, was something I was never especially happy
with, as whilst estimating pre-allocation for the string could've
improved performance, I believed using `collect()` would produce a
cleaner implementation. Though, unable to devise a cleaner solution due
to effectively requiring memory to be allocated for every item whether
processed or not, or otherwise incur a type conflict, I did not venture
to update the implementation, until I encountered the `Cow` smart
pointer...

Due to entirely solving the type conclict, this updates the
implementation of `pigify()` to map the words of `convert` to
`Cow` strings, and `collect()` them, including reverting the method
return type to it's original signiture, rather than returning a result.

BREAKING CHANGE: Function signiture is changed to simply return a
`String`, rather than a `Result`, due to no longer implementing an error
case.